rather than just `derive`ing eq/hash, impl to wrap `val`'s
otherwise fucking everything needs to be Hash and Eq, which doesn't make
sense. I don't think this can cause data integrity issues, but I'll have
to rethink it when the guy sitting behind me finishes his boxing match
with my seat. I swear to christ he's like an epileptic cat at a fish
tank but he was abandoned at birth and was raised by wild jackhammers

also may need to rethink the `Rc` model (or size?) because it can't be
mutated if there are >1 references. Which there almost certainly will be
always. Hmmm.