make Pair types only generic over those that pair with them
Not sure how to properly store edges. Edges are guaranteed to be
trivially copyable (they contain 2 `Addr`s, and maybe a weight. So it's
not that big of a deal to duplicate edges (e.g. store one in its
parent and one in its child). But not sure it's ideal.

Instead it might be cool to use another `AddrHashSet` to store them.
That would mean there's just a different `Copy`able type used to refer
to them and that edge weights could be generic (signed vs unsigned,
arbitrary precision, funky stuff, etc).