begin switching to `Addr` to refer to everything
Just referencing everything by `Addr` is a little too unsafe for me. Not
sure the best way to fix this.

Having the `Signature` in debug mode is a start. That imposes a runtime
check that an `Addr` can't be used in the wrong collection. But still.

I might add a `PhantomData` to `Addr` and give it the type of whatever
it associates with. That would complicate some things (now vertices need
to know the types of their edges and edges their vertices) but I think
it's a worthwhile sacrifice (and would probably make some things
clearer).

Maybe add a reference to the Map itself to the Addr type to ensure the
Map can't outlive it? That seems hard because the map can't be modified.
hmmm.