switch to Rc instead of Box and unsafe lifetime hack
I don't love the `Rc` option because I don't think there's anything
wrong with the ownership hierarchy previously in the graph: each
`Vertex` owns its `V`, the `Graph` owns each `Edge`, etc. The `Rc` model
doesn't preserve this, instead treating an `Rc<V>` the same whether it's
in a `Vertex` or an `Edge`. But for now at least (and even for later I
think) this is preferable. There's basically no runtime/memory overhead
and it makes things cleaner and safer.

One thing to note is that this isn't `Rc<Refcell<T>>`. The contents of
an `Rc` can't be mutated; therefore edges and `V`s are immutable. This
is fine, but it's the reason we refer to vertices by `Rc<V>` instead of
`Rc<Vertex`; because adding neighbors requires mutability.
I think this is okay. It would be kind of convenient for an edge to
refer directly to a vertex like it would in C/C++. But to have even an
inkling of ownership, this has to change. I think it makes for an `Edge`
to "know" the values of its endpoints but nothing else; on its own it
"knows" nothing of the graph hierarchy, just its part.
The head and tail are properties of an `Edge`, but not its nodes'
neighbors. Neighbor edges are properties of a `Vertex`, but not what's
on the other endpoint.
If you want to learn anything else, you require a ref to the `Graph`,
which is the owner of the actual global hierarchy. We still need `Rc`s
because data is shared between vertices / edges and vertices / edges /
everything and the `Graph`, but I like this option.

I suspect this makes sense. Time will tell.