reworking edge layout, how generic impls are
Instead of an `Edge` storing both a lhs and rhs, it only stores an
"opposite end", which makes it easier to deal with edges generically.
This makes it possible to make the DF/BF/neighbor/etc. traversals to be
generic over all `Graph`s. This isn't actually necessary I don't think,
but it will make some things easier in the future I think (fewer
unnecessary case-by-case handling situations).

concern: I think edges can appear in a `hood` twice? Not 100% certain.
Might just be cleaner to use a HashMap of neighbors to edge values
(which would be cool but higher overhead)